fix: 빌드 에러 수정

+++ src/app/home/setting/expenseCategory/_components/DeleteExpenseCategoryModal/DeleteExpenseCategoryModal.hook.tsx
@@ -14,6 +14,8 @@ export default function useDeleteExpenseCategoryModal({
 
     const handleDeleteButtonClick = () => {
         // 낙관적 업데이트
+        if (!expenseCategoriesDispatch) return;
+
         expenseCategoriesDispatch({
             type: "DELETE",
             payload: {
